+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Modify interface for functions `optional` and `cluster`.
 - Replace generate for blocks with elaborated clusters.
 - Move cluster elaboration to evaluate stage.
+- Support `Wire` ports for bidirectional signals.
 
 ## [0.1.14]
 ### Importer